add ChatActionableButtons
